requirements.txt
scikit-allel==1.3.6

numpy==1.24.0
pandas==1.5.2
scipy==1.9.3
parmed==3.4.3
matplotlib==3.6.2
venn==0.1.3

plotly==5.11.0
flask_caching==2.0.1

dash==2.18.2
dash_bio==1.0.2
dash_bootstrap_components==1.2.1
dash_bootstrap_templates==1.0.7
feffery-antd-components==0.3.11
